Remove Hitscan from library Protocol
